Working web app & Basically Finished Discord Bot
The web app works perfectly. Pulling all of the songs from uri.txt and adding them to a set spotify playlist.

The next big hurdle to get over is to have the discord bot make a request to localhost:5000 to create that request and once a code 200 is recived. The bot needs to clear the URI.txt files so the playlist dosent keep getting duplicate songs added to it.