Fixed grabpast and added a todo list in readme
Grab past has been fixed, the flag will now go off preventing the function to be ran more than once. To send off the songs to the playlist, you must send one new song to add all of the songs to the playlist. Either that or make an enritrly seprate function to send off the songs.