Removed out of date / annoying features
Removed app.py's ability to add songs to a playlist. This is only a function playlist_update.py should have.

In spotbot.py no longer replies to a users message when a song is added to the set Spotify playlist. A check mark emoji is added to the users message only now as the users main way of receiving feedback.